quant-research-v1/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quant-research-v1"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # Core
    "duckdb>=1.0.0",
    "polars>=0.20.0",
    "pyarrow>=15.0.0",
    # Data ingestion
    "httpx>=0.27.0",
    "yfinance>=0.2.40",
    "pandas-datareader>=0.10.0",
    "fredapi>=0.5.0",
    "exchange-calendars>=4.5.0",
    "tenacity>=8.2.0",
    # Config / CLI
    "typer>=0.12.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.0",
    "structlog>=24.0.0",
    "orjson>=3.10.0",
    # Analytics
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "statsmodels>=0.14.0",
    "scikit-learn>=1.4.0",
    "cvxpy>=1.4.0",
    "hmmlearn>=0.3.0",
    # Reporting
    "jinja2>=3.1.0",
    "matplotlib>=3.8.0",
    "markdown>=3.6",
    "anthropic>=0.28.0",
    "premailer>=3.10.0",
    # Gmail delivery
    "google-api-python-client>=2.120.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "hypothesis>=6.100.0",
    "respx>=0.21.0",
    "freezegun>=1.4.0",
    "ruff>=0.4.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
